refactored seed scripts
