+++ script/finalFoursquareData.json
@@ -52,7 +52,6 @@
   {"name": "Doughnut Plant", "rating": 9.4},
   {"name": "Tanoshi Sushi", "rating": 9.4},
   {"name": "Bobwhite Counter", "rating": 9.4},
-  {"name": "Cocoron", "rating": 9.4},
   {"name": "Miss Lily's 7A", "rating": 9.3},
   {"name": "Housing Works Bookstore Cafe", "rating": 9.3},
   {"name": "dell'anima", "rating": 9.3},
