+++ script/googleTestScript.js
@@ -6,27 +6,34 @@ const utf8 = require('utf8')
 
 const iconUrl = 'https://image.flaticon.com/icons/svg/281/281781.svg'
 
-async function fetchRestaurants() {
+async function googleCreate(createDbRestaurantObj) {
   try {
-    const restaurants = await Restaurant.findAll()
-    return restaurants
-  } catch (error) {
-    console.error(error)
-  }
-}
-
-async function createDbRestaurantObj() {
-  try {
-    const restaurants = await fetchRestaurants()
-    let restaurantObj = {}
-
-    restaurants.forEach(restaurant => {
-      let tempRestaurant = restaurant.dataValues
-      if (!restaurantObj[tempRestaurant.name]) {
-        restaurantObj[tempRestaurant.name] = tempRestaurant
-      }
+    console.log('Creating Google reviews...')
+    const googleData = JSON.parse(
+      fs.readFileSync('script/googleData.json', 'utf8')
+    )
+    const newGoogleData = googleData.filter(restaurant => {
+      if (restaurant) return true
     })
-    return restaurantObj
+
+    const restaurantDbObj = await createDbRestaurantObj()
+
+    const bulkCreateArr = newGoogleData
+      .filter(restaurant => {
+        if (restaurantDbObj[restaurant.name]) {
+          return true
+        }
+      })
+      .map(restaurant => {
+        return {
+          source: 'Google',
+          rating: restaurant.rating,
+          restaurantId: restaurantDbObj[restaurant.name].id,
+          sourceLogo: iconUrl
+        }
+      })
+    await Review.bulkCreate(bulkCreateArr)
+    console.log('Google is done!')
   } catch (error) {
     console.error(error)
   }
@@ -62,38 +69,5 @@ async function createDbRestaurantObj() {
 //     console.error(error)
 //   }
 // }
-
-async function googleCreate() {
-  try {
-    console.log('Creating Google reviews...')
-    const googleData = JSON.parse(
-      fs.readFileSync('script/googleData.json', 'utf8')
-    )
-    const newGoogleData = googleData.filter(restaurant => {
-      if (restaurant) return true
-    })
-
-    const restaurantDbObj = await createDbRestaurantObj()
-
-    const bulkCreateArr = newGoogleData
-      .filter(restaurant => {
-        if (restaurantDbObj[restaurant.name]) {
-          return true
-        }
-      })
-      .map(restaurant => {
-        return {
-          source: 'Google',
-          rating: restaurant.rating,
-          restaurantId: restaurantDbObj[restaurant.name].id,
-          sourceLogo: iconUrl
-        }
-      })
-    await Review.bulkCreate(bulkCreateArr)
-    console.log('Done!')
-  } catch (error) {
-    console.error(error)
-  }
-}
 
 module.exports = googleCreate
